new_graphic loads velocity and angle file now. Backprop works! Also evolutionary algorithms look better!

+++ logic/include/evolutionary.hpp
@@ -69,8 +69,8 @@ class evolutionary
 		
 		bool datasetwritten;
         //double max_delta; 
-        std::vector< std::vector<float> > trainingdataV;
-        std::vector< std::vector<float> > trainingdataA;
+        std::vector< std::vector<double> > trainingdataV;
+        std::vector< std::vector<double> > trainingdataA;
 
         double revert_agent[100][100][100];
 
@@ -93,8 +93,8 @@ class evolutionary
 
         int get_bestFitness_overall();
 
-        void set_trainingdataV(std::vector< std::vector<float> > ptrainingdata);
-        void set_trainingdataA(std::vector< std::vector<float> > ptrainingdata);
+        void set_trainingdataV(std::vector< std::vector<double> > ptrainingdata);
+        void set_trainingdataA(std::vector< std::vector<double> > ptrainingdata);
 
     private:
